LearnLevelHub/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "openai>=1.72.0",
    "streamlit>=1.44.1",
]
